SCHED-615: Nearly done.
