+++ scheduler/core/builder/modes.py
@@ -2,6 +2,7 @@
 # For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
 from enum import Enum
+from typing import final
 
 import strawberry  # noqa
 
@@ -11,6 +12,13 @@ from .validationbuilder import ValidationBuilder
 from ..eventsqueue import EventQueue
 
 
+__all__ = [
+    'SchedulerModes',
+    'dispatch_with',
+]
+
+
+@final
 @strawberry.enum
 class SchedulerModes(Enum):
     """Scheduler modes available:
