+++ scheduler/core/components/base/__init__.py
@@ -1,11 +1,4 @@
 # Copyright (c) 2016-2024 Association of Universities for Research in Astronomy, Inc. (AURA)
 # For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-from abc import ABC
-
-
-class SchedulerComponent(ABC):
-    """
-    Base class for all Scheduler components involved in the pipeline.
-    """
-    ...
+from .schedulercomponent import *
