+++ scheduler/core/components/base/schedulercomponent.py
@@ -1,0 +1,16 @@
+# Copyright (c) 2016-2024 Association of Universities for Research in Astronomy, Inc. (AURA)
+# For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
+
+from abc import ABC
+
+
+__all__ = [
+    'SchedulerComponent',
+]
+
+
+class SchedulerComponent(ABC):
+    """
+    Base class for all Scheduler components involved in the pipeline.
+    """
+    ...
