+++ scheduler/core/components/optimizer/timeline.py
@@ -3,7 +3,7 @@
 
 from dataclasses import dataclass
 from datetime import datetime, timedelta, timezone
-from typing import ClassVar, List, Mapping, Optional, Sequence, Tuple
+from typing import final, ClassVar, List, Mapping, Optional, Sequence, Tuple
 
 import numpy as np
 from lucupy.minimodel import NightIndex, Observation, ObservationID, Site
@@ -12,6 +12,13 @@ from lucupy.types import Interval, ZeroTime
 from scheduler.core.calculations.nightevents import NightEvents
 
 
+__all__ = [
+    'Timeline',
+
+]
+
+
+@final
 @dataclass
 class Timeline:
     """
