+++ scheduler/core/eventsqueue/__init__.py
@@ -1,5 +1,5 @@
 # Copyright (c) 2016-2024 Association of Universities for Research in Astronomy, Inc. (AURA)
 # For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-from .event_queue import *
+from .eventqueue import *
 from .events import *
