+++ scheduler/core/eventsqueue/events.py
@@ -11,6 +11,24 @@ from lucupy.minimodel import Resource, TimeslotIndex, VariantChange
 from lucupy.timeutils import time2slots
 
 
+__all__ = [
+    'UUIDIdentified',
+    'UUIDReferenced',
+    'Event',
+    'RoutineEvent',
+    'TwilightEvent',
+    'EveningTwilightEvent',
+    'MorningTwilightEvent',
+    'InterruptionEvent',
+    'WeatherChangeEvent',
+    'FaultEvent',
+    'EngineeringTaskEvent',
+    'InterruptionResolutionEvent',
+    'FaultResolutionEvent',
+    'EngineeringTaskResolutionEvent',
+]
+
+
 @dataclass
 class UUIDIdentified(ABC):
     """
