+++ scheduler/core/meta/__init__.py
@@ -1,6 +1,12 @@
 # Copyright (c) 2016-2024 Association of Universities for Research in Astronomy, Inc. (AURA)
 # For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
+
+__all__ = [
+    'Singleton',
+]
+
+
 class Singleton(type):
     """
     Thread-safe Singleton metaclass.
