+++ scheduler/core/output/__init__.py
@@ -23,6 +23,18 @@ from scheduler.core.programprovider.ocs import OcsProgramProvider
 from scheduler.core.calculations.selection import Selection
 
 
+__all__ = [
+    'print_program_from_provider',
+    'print_collector_info',
+    'print_atoms_for_observation',
+    'atoms_to_sheet',
+    'print_plans',
+    'plans_table',
+    'pickle_plans',
+    'pickle_selection',
+]
+
+
 def print_program_from_provider(filename=os.path.join('data', 'GN-2018B-Q-101.json.gz'),
                                 provider: ProgramProvider = OcsProgramProvider) -> None:
     """
