+++ scheduler/core/service/__init__.py
@@ -1,4 +1,4 @@
 # Copyright (c) 2016-2024 Association of Universities for Research in Astronomy, Inc. (AURA)
 # For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-from .service import Service
+from .service import *
